server/requirements.txt
Django==2.2.14
psycopg2-binary==2.8.5
gunicorn==20.0.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
ibm-watson==5.1.0
ibm-cloud-sdk-core==3.10.0
boto3==1.17.58
django-storages==1.11.1
Pillow==8.2.0